requirements.txt
opencv-python>=4.7.0
opencv-contrib-python>=4.7.0
numpy>=1.23.0
Pillow>=9.0.0
